Fix testfx jigsaw module reflection error
TestFX throws runtime errors because it's unable to get access to javafx.graphics modules, since it doesn't support Java modules. This fix allows access to those internal modules to TestFX. It's hoped to be fixed in TestFX, so this commit should be reverted as well if that happens: https://github.com/TestFX/TestFX/issues/638.